technical-analysis/requirements.txt
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
parsedatetime==2.4
python-telegram-bot==13.3
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
seaborn==0.11.0
ta==0.7.0
tornado==6.1
urllib3==1.25.8
yahoo-fin==0.8.7
yfinance==0.1.55
alpaca-trade-api==0.53.0
alpha-vantage==2.3.1
robin-stocks==1.6.2